# [libc++] Optimize num_put integral functions