@nico 2025-03-05
We have a test that checks that streaming a pointer to an ostringstream prints a `0x` prefix on non-win but not on win.

(I suppose the test is just supposed to document current behavior.)

This seems to change the behavior on windows to print a `0x` prefix there too. Is that intentional?