@philnik777 2025-03-05
> We have a test that checks that streaming a pointer to an ostringstream prints a `0x` prefix on non-win but not on win.
> 
> (I suppose the test is just supposed to document current behavior.)
> 
> This seems to change the behavior on windows to print a `0x` prefix there too. Is that intentional?

Depends on what you mean by "intentional". We're aware that this changed and think that the improved performance is worth the change in behaviour. It also makes the output more consistent across platforms. Do you think this is important enough that we should try to mimic the platforms behaviour?

> Also, one of these now seems to fail on chromeos: https://source.chromium.org/chromium/chromium/src/+/main:third_party/shell-encryption/src/int256_test.cc;l=527

Since `int256` seems to be a custom type of yours, I'm not sure how it outputs stuff. Could you give a bit more detail on what change you observe?