@philnik777 2025-03-05
> > Depends on what you mean by "intentional". We're aware that this changed and think that the improved performance is worth the change in behaviour. It also makes the output more consistent across platforms. Do you think this is important enough that we should try to mimic the platforms behaviour?
> 
> No, I was just wondering if it's intentional, since the commit description says "optimize" and normally optimizations don't change behavior, and it added a libcxx/docs/ReleaseNotes/20.rst change but didn't mention the behavior change. We can update the test if it's intentional.

Yeah, maybe we should add a release notes about that. Seems like it's worth mentioning.