@alexfh 2025-03-14
> > No, I was just wondering if it's intentional, since the commit description says "optimize" and normally optimizations don't change behavior, and it added a libcxx/docs/ReleaseNotes/20.rst change but didn't mention the behavior change. We can update the test if it's intentional.
> 
> Yeah, maybe we should add a release notes about that. Seems like it's worth mentioning.

It's definitely worth adding release notes, especially given that the new behavior diverges both from the previous libc++ and the current libstdc++ behavior - https://gcc.godbolt.org/z/8szE6997z. There's also a change in how nullptr is formatted in some cases, but I haven't yet figured out a specific case.