@alexfh 2025-03-14
Actually, this behavior change affects a non-trivial amount of code, both Google-internal and open-source (for now, I found breakages in Abseil and tinyformat - https://github.com/c42f/tinyformat). I wonder if the C++ standard says anything about the particular formatting that should be used in these cases? https://gcc.godbolt.org/z/ba8bEzj99