@alexfh 2025-03-14
I found https://bugzilla.redhat.com/show_bug.cgi?id=166735, which claims that 0x0 should just be 0, IIUC.