@alexfh 2025-03-14
The relevant parts of the recent C++ and C standards are: https://eel.is/c++draft/facet.num.put.virtuals, https://eel.is/c++draft/intro.refs, and https://www.open-std.org/jtc1/sc22/wg14/www/docs/n2310.pdf 7.21.6.1 paragraph 6 (page 229).